chore: refactor _send_to_audit_protocol > Callback Worker class, minor bug fix
